+++ src/views/UserEdit/UserForm.js
@@ -110,7 +110,9 @@ class UserForm extends React.Component {
     onCancelKeycloakConfirmation: PropTypes.func,
     confirmCreateKeycloakUser: PropTypes.func,
     setAssignedRoleIds: PropTypes.func,
-    assignedRoleIds: PropTypes.arrayOf(PropTypes.string),
+    assignedRoleIds: PropTypes.object,
+    setTenantId: PropTypes.func,
+    tenantId: PropTypes.string
   };
 
   static defaultProps = {
@@ -475,6 +477,9 @@ class UserForm extends React.Component {
                           setButtonRef={this.setButtonRef}
                         /> : <EditUserRoles
                           form={form}
+                          user={initialValues}
+                          setTenantId={this.props.setTenantId}
+                          tenantId={this.props.tenantId}
                           setAssignedRoleIds={this.props.setAssignedRoleIds}
                           assignedRoleIds={this.props.assignedRoleIds}
                           accordionId="userRoles"